apps/api/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "roleiq-api"
version = "0.1.0"
description = "Backend API for RoleIQ"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.13.2",
  "fastapi>=0.115.0",
  "httpx>=0.27.2",
  "openai>=1.54.0",
  "pgvector>=0.3.5",
  "psycopg[binary]>=3.2.1",
  "pydantic-settings>=2.5.2",
  "PyMuPDF>=1.24.10",
  "python-dateutil>=2.9.0.post0",
  "python-docx>=1.1.2",
  "python-multipart>=0.0.9",
  "SQLAlchemy>=2.0.35",
  "tenacity>=9.0.0",
  "uvicorn[standard]>=0.31.0",
]

[project.optional-dependencies]
dev = [
  "mypy>=1.12.0",
  "pytest>=8.3.3",
  "pytest-asyncio>=0.24.0",
  "pytest-cov>=5.0.0",
  "ruff>=0.6.8",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"
exclude = [".venv", ".uv-cache"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["B008", "E501"]

[tool.mypy]
python_version = "3.12"
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
strict_optional = true
ignore_missing_imports = true
